requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
colorzero==2.0
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.2
Flask-SocketIO==5.3.6
flatbuffers==20181003210633
fonttools==4.47.2
future==1.0.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gpiod==2.1.3
gpiozero==2.0.1
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
idna==3.6
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.15.0
kiwisolver==1.4.5
lgpio==0.0.0.2
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
ml-dtypes==0.2.0
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==2.0.0
pigpio==1.78
pillow==10.2.0
ply==3.11
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pyomo==6.7.1
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tflite-runtime==2.14.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0